+++ pkg/fanal/analyzer/imgconf/apk/apk_test.go
