+++ pkg/fanal/analyzer/imgconf/apk/testdata/history_v3.7.json
