+++ pkg/fanal/analyzer/imgconf/apk/testdata/history_v3.9.json
